unify test files
